requirements.txt
numpy
pandas
Pillow
opencv-python
pycocotools  
torch
torchvision
pytorch_lightning
torchmetrics
matplotlib
